+++ lib/integral.rb
@@ -16,9 +16,8 @@ require 'integral/grids/posts_grid'
 require 'integral/grids/images_grid'
 require 'integral/acts_as_listable'
 require 'integral/acts_as_integral'
-require 'integral/widgets/recent_posts'
-require 'integral/widgets/swiper_list'
-require 'integral/content_renderer'
+require 'integral/blocks/recent_posts'
+require 'integral/block_list_renderer'
 require 'integral/list_renderer'
 require 'integral/swiper_list_renderer'
 require 'integral/list_item_renderer'
@@ -56,9 +55,6 @@ module Integral
 
   mattr_accessor :additional_settings_params
   @@additional_settings_params = []
-
-  mattr_accessor :additional_widgets
-  @@additional_widgets = []
 
   mattr_accessor :additional_post_params
   @@additional_post_params = []
@@ -144,5 +140,12 @@ module Integral
   # @return [Boolean] Enables Dynamic Routing of the homepage using Integral::Middleware::Router
   def self.dynamic_homepage_enabled?
     Integral.root_path.nil?
+  end
+
+  # @return [Array] Dynamic block editor block types
+  def self.dynamic_blocks
+    # TODO: Allow host app to add additional blocks
+    # i.e. blocks.concat Integral.additional_dynamic_blocks
+    [ Integral::Blocks::RecentPosts ]
   end
 end